fix: images not loading on github pages
The image URLs in the markdown files were not loading correctly on the GitHub Pages site. This issue arose due to the URLs directing to the "blob" directory in the GitHub repository. The present commit resolves this problem by adjusting the URLs to point to the "raw" directory instead.

The regex used to find all the images that were being loaded from the blob directory was:

```
\/blob\/(?=.*\.(png|jpg|jpeg|gif|svg|bmp|webp))
```